pyproject.toml
[tool.poetry]
name = "news-compare"
version = "0.1.0"
description = ""
authors = ["Samu Syrjänen <samu.syrjanen@gmail.com>"]
readme = "README.md"
packages = [{include = "news_compare"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
